% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{iowamodels-package}
\alias{iowamodels}
\alias{iowamodels-package}
\title{iowamodels: Modular reinforcement learning models for the Iowa gambling task}
\description{
cmdstanr backend for the iowa package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://areshenk-rpackages.github.io/iowamodels/}
}

}
\concept{help}
